Validation through Redmine
In addition to allowing an entire domain, or using a file of allowed
email addresses, support looking up users in Redmine.

A Redmine URL, API key, and group ID are passed as arguments.  The
Redmine API will be used to verify that that user is in the specified
group for validation.